Move nanoflann.h to canonical location.
